+++ src/index.ts
@@ -112,9 +112,8 @@ class Vditor {
 
             const ui = new Ui(this.vditor);
 
-
-           if (mergedOptions.after) {
-               mergedOptions.after()
+            if (mergedOptions.after) {
+               mergedOptions.after();
            }
         });
     }