+++ CHANGELOG.md
@@ -20,8 +20,8 @@
 
 ### v1.9.1 / 2019-11-04
 
-* [157](https://github.com/b3log/vditor/issues/157) 添加 options.after `feature`
-* [156](https://github.com/b3log/vditor/issues/156) 语法高亮添加行号 `feature`
+* [157](https://github.com/b3log/vditor/issues/157) 初始化后添加回调方法 `options.after` `feature`
+* [156](https://github.com/b3log/vditor/issues/156) 语法高亮添加行号配置 `options.preview.hljs.lineNumber` `feature`
 * [155](https://github.com/b3log/vditor/issues/155) preview 方法支持多次渲染 `enhancement`
 * [153](https://github.com/b3log/vditor/issues/153) Markdown 渲染空格问题 `bug`
 * [152](https://github.com/b3log/vditor/issues/152) 直接调用 getHTML 抛错 `bug`